@@ -24,7 +24,7 @@
     "deploy": "gh-pages -d example/build"
   },
   "peerDependencies": {
-    "react": "^17.0.2"
+    "react": "^18.0.2"
   },
   "devDependencies": {
     "@babel/parser": "^7.24.5",
@@ -35,15 +35,15 @@
     "@types/jest": "^25.1.4",
     "@types/lodash": "^4.17.4",
     "@types/node": "^12.12.38",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.0.2",
+    "@types/react-dom": "^18.0.2",
     "axios": "^1.7.2",
     "gh-pages": "^2.2.0",
     "microbundle-crl": "^0.13.11",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-scripts": "^3.4.1",
     "typescript": "^3.7.5"
   },
