@@ -1,23 +1,25 @@
-import { CacheRequest,useFormModal } from 'avc-hooks'
+import { CacheRequest } from 'avc-hooks'
 import 'avc-hooks/dist/index.css'
 import React, { useEffect } from 'react'
 import axios from 'axios'
-import { Button, Form, Input, InputNumber } from 'antd'
+import { Button } from 'antd'
+import useModal from './useModal'
+
+
 const cacheAxios = CacheRequest(axios,{
     updateKey:"updateKey", // 在请求中添加updateKey参数，当updateKey为true时，请求不会使用缓存，而是重新请求 默认：false
     cacheStore:"cacheStore" // 存储库名，默认：avc-cache
 })
-cacheAxios.interceptors.response.use(response=>{
+cacheAxios.interceptors.response.use((response: { data: any })=>{
     return response.data
 })
-
-interface FormValues {
-    id?: number;
-    name: string;
-    age: number;
+export {
+    cacheAxios
 }
 
+
 const App = () => {
+
     useEffect(() => {
         console.time('GET缓存请求')
         cacheAxios.get('http://127.0.0.1:8080/app', {
@@ -39,59 +41,18 @@ const App = () => {
             console.timeEnd('GET不读缓存')
             console.log("GET不缓存结果",res)
         })
-        console.time('POST请求')
-        cacheAxios.post('http://127.0.0.1:8080/data', {}).then((res: any) => {
-            console.timeEnd('POST请求')
-            console.log("POST结果",res)
-        })
+
     }, [])
-    const content = <>
-        <Form.Item
-            label={"姓名"}
-            name={"name"}
-        >
-            <Input placeholder={"请输入姓名"} />
-        </Form.Item>
-        <Form.Item
-            label={"年龄"}
-            name={"age"}
-        >
-            <InputNumber placeholder={"请输入年龄"} />
-        </Form.Item>
-    </>
-    const testFormModal = useFormModal({
-        title: '创建弹窗',
-        content,
-        onOk(values: FormValues){
-            return cacheAxios.post('http://127.0.0.1:8080/data', values).then(() => {
-                return {
-                    status: true,
-                    message: "创建成功"
-                }
-            })
-        }
-    },[]);
-    const updateModal = useFormModal({
-        title: '编辑弹窗',
-        content,
-        onOk(values:FormValues){
-            return cacheAxios.post('http://127.0.0.1:8080/data', values).then(() => {
-                return {
-                    status: true,
-                    message: "修改成功"
-                }
-            }).catch(() => {
-                return {
-                    status: false,
-                    message: "修改失败"
-                }
-            })
-        }
-    })
+
+    const {
+        createModal,
+        updateModal
+    } = useModal();
+
     return <>
         <Button
             onClick={() => {
-                testFormModal.setOpen(true)
+                createModal.setOpen(true)
             }}
         >新增</Button>
         <Button
